Cleanup utilities
